Hot Fix: db_sum() Error Handling and Fixes Moved the creation of tmp_path inside the try except block, finally, checking if that path exists before deleting the tree.
Also the local hash db is no longer recreated, instead uses local one already in .pyupdate folder

+++ pyupdate/update.py
@@ -28,7 +28,7 @@ class UpdateManager:
         self._project_path = project_path
         self._pyupdate_path = os.path.join(self._project_path, '.pyupdate')
         self._config_path = os.path.join(self._pyupdate_path, 'config.yaml')
-        self._hash_db_path = None  # Set in _validate_attributes
+        self._local_hash_db_path = None  # Set in _validate_attributes
 
         self._config_man = helper.Config()
         self._web_man = None  # Set in _validate_attributes
@@ -54,7 +54,7 @@ class UpdateManager:
         self._project_path = value
         self._pyupdate_path = os.path.join(self._project_path, '.pyupdate')
         self._config_path = os.path.join(self._pyupdate_path, 'config.yml')
-        self._hash_db_path = None  # Set in _validate_attributes
+        self._local_hash_db_path = None  # Set in _validate_attributes
         self._validate_attributes()
     
     def _validate_attributes(self) -> None:
@@ -71,11 +71,11 @@ class UpdateManager:
             raise FileNotFoundError(self._config_path)
         
         config_data = self._config_man.load_yaml(self._config_path)
-        self._hash_db_path = os.path.join(self._pyupdate_path, config_data['hash_db'])
+        self._local_hash_db_path = os.path.join(self._pyupdate_path, config_data['hash_db'])
         self._web_man = helper.Web(self._url)
 
-        if not os.path.exists(self._hash_db_path):
-            raise FileNotFoundError(self._hash_db_path)
+        if not os.path.exists(self._local_hash_db_path):
+            raise FileNotFoundError(self._local_hash_db_path)
 
     def check_update(self) -> dict:
         """Compare cloud and local version and return a dict with the results"""
@@ -94,18 +94,16 @@ class UpdateManager:
     
     def db_sum(self) -> hashing.DBSummary:
         """Return a DBSummary object using the cloud and local hash databases"""
-        tmp_path = tempfile.mkdtemp()
-
+        tmp_path = ""
         try:
+            tmp_path = tempfile.mkdtemp()
             hasher = hashing.Hasher(project_name=os.path.basename(self._project_path))
 
-            cloud_hash_db = self._web_man.download_hash_db(os.path.join(tmp_path, 'cloud_hashes.db'))
-            local_hash_db = hasher.create_hash_db(self._project_path,
-                                                os.path.join(tmp_path, 'local_hashes.db'),
-                                                exclude_paths=[self._pyupdate_path])
+            cloud_hash_db_path = self._web_man.download_hash_db(os.path.join(tmp_path, 'cloud_hashes.db'))
 
-            return hasher.compare_databases(local_hash_db, cloud_hash_db)
+            return hasher.compare_databases(self._local_hash_db_path, cloud_hash_db_path)
         except Exception as error:
             raise error
         finally:
-            shutil.rmtree(tmp_path)
+            if os.path.exists(tmp_path):
+                shutil.rmtree(tmp_path)
